Separate model from services namespace
