+++ Sources/Subscription/Flows/AppStore/AppStoreRestoreFlow.swift
@@ -79,8 +79,8 @@ public final class AppStoreRestoreFlow {
         var isSubscriptionActive = false
 
         switch await SubscriptionService.getSubscriptionDetails(token: accessToken) {
-        case .success(let response):
-            isSubscriptionActive = response.isSubscriptionActive
+        case .success(let subscription):
+            isSubscriptionActive = subscription.isActive
         case .failure:
             os_log(.error, log: .subscription, "[AppStoreRestoreFlow] Error: failedToFetchSubscriptionDetails")
             return .failure(.failedToFetchSubscriptionDetails)
